rm observer pattern for simplicity
